requirements.txt
httpx==0.24.1
pyyaml==6.0
numpy==1.24.3
sentence-transformers==2.2.2
pydantic==2.0.2
click==8.1.3
python-dotenv==1.0.0
